Add flattenIterator().
